+++ Examples/YandexMobileAdsExample/YandexMobileAdsExample/Accessibility/GDPRAccessibility.swift
@@ -1,3 +1,10 @@
+/*
+ * Version for iOS © 2015–2025 YANDEX
+ *
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at https://yandex.com/legal/mobileads_sdk_agreement/
+ */
+
 enum GDPRAccessibility {
     static let settingsButton = "gdpr_settings"
     static let acceptButton = "gdpr_accept"