+++ Examples/YandexMobileAdsExample/YandexMobileAdsExample/YandexAdsExampleViewController.swift
@@ -1,9 +1,10 @@
 /*
- * Version for iOS © 2015–2023 YANDEX
+ * Version for iOS © 2015–2025 YANDEX
  *
  * You may not use this file except in compliance with the License.
  * You may obtain a copy of the License at https://yandex.com/legal/mobileads_sdk_agreement/
  */
+ 
 import Foundation
 import UIKit