+++ Examples/YandexMobileAdsExample/YandexMobileAdsExampleUITests/Pages/RootPage.swift
@@ -1,3 +1,10 @@
+/*
+ * Version for iOS © 2015–2025 YANDEX
+ *
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at https://yandex.com/legal/mobileads_sdk_agreement/
+ */
+
 import XCTest
 
 struct RootPage: PageObject {
@@ -5,7 +12,6 @@ struct RootPage: PageObject {
 
     var yandex: XCUIElement { element(id: Keys.yandex, type: .cell) }
     var mobileMediation: XCUIElement { element(id: Keys.mobileMediation, type: .cell) }
-    var thirdPartyMediation: XCUIElement { element(id: Keys.thirdPartyMediation, type: .cell) }
     var gdpr: XCUIElement { element(id: Keys.gdpr, type: .cell) }
     var adFox: XCUIElement { element(id: Keys.adFox, type: .cell) }