+++ changelog/mediation/mobileads-chartboost/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 9.8.1.4
+
+### Updated
+
+- Updated minimum supported `YandexMobileAds` version to `7.15.0`
+
 ## Version 9.8.1.3
 
 ### Updated